+++ Common/ERP.Entities/Models/ProductEntities/ProductSKU.cs
@@ -13,7 +13,7 @@ namespace ERP.Entities.Models
 
         public int ID { get; set; }
         [ForeignKey("Product")]
-        public Nullable<int> ProductId { get; set; }
+        public Nullable<long> ProductId { get; set; }
         public virtual Product Product { get; set; }
         public string SKUSize { get; set; }
         public float MRP { get; set; }
